Implement the 1-call unary function wrapper class
... as well as its static factory method.
The callback null-check is performed in the latter, rather than in the
constructor, to avoid the object allocation and constructor call cost if
it is not necessary (minor optimization), and to simplify the
constructor body: it now only initializes the object's field. This is in
my view cleaner code (better separation of concerns).